fix: small fixes  - ensure testdata files end in newlines  - add full stop to comments  - move sumTypeDecl to new file decl.go  - remove unnecessary `return` statement  - improve regex in TestNotSealed
