+++ pkg/sumtype/testdata/src/p/notsealed.go
@@ -3,4 +3,4 @@ package main
 //go-sumtype:decl NotSealedT
 
 // TestNotSealed
-type NotSealedT interface {} // want "not sealed"
+type NotSealedT interface{} // want "interface 'NotSealedT' is not sealed"
